fix code style and adjust linting (1h)
Consider switching to eslint. I am not even sure what exactly style.lint is for. Might be an ancient thingy. In order to work properly together we should align on linting tools and code style. I prefer to stay in-sync with the existing naming conventions for filenames for instance. It is common to use camelCase or PascalCase. File system should be clean and easy to read. What to you thing?